+++ tools/unsteady.go
@@ -1,0 +1,9 @@
+package tools
+
+// Unsteady Data
+type Unsteady struct {
+	InitialConditions  interface{} // to be implemented
+	BoundaryConditions interface{}
+	MeterologicalData  interface{} // to be implemented
+	ObservedData       interface{} // to be implemented // added in version 6.2
+}
